requirements.txt
Django>=4.2
whitenoise
gunicorn
djangorestframework
psycopg2
celery>=5.3
redis>=5.0 
channels
channels_redis
daphne
dotenv